Feat/#86:Submission Form 연관관계 구현 및 권한검증 추가

+++ cluvr-api/src/main/java/com/example/cluvrapi/domain/applicationForm/service/SubmissionFormServiceImpl.java
@@ -15,7 +15,10 @@ import com.example.cluvrapi.domain.applicationForm.repository.SubmissionFormRepo
 import com.example.cluvrapi.domain.club.entity.Club;
 import com.example.cluvrapi.domain.club.enums.JoinType;
 import com.example.cluvrapi.domain.club.repository.ClubRepository;
+import com.example.cluvrapi.domain.clubMember.entity.ClubMember;
+import com.example.cluvrapi.domain.clubMember.repository.ClubMemberRepository;
 import com.example.cluvrapi.domain.common.dto.PageResponseDto;
+import com.example.cluvrapi.domain.common.validator.ClubValidator;
 import com.example.cluvrapi.global.exception.BusinessException;
 import com.example.cluvrapi.global.response.ResponseCode;
 
@@ -25,28 +28,44 @@ public class SubmissionFormServiceImpl implements SubmissionFormService {
 
 	private final SubmissionFormRepository submissionFormRepository;
 	private final ClubRepository clubRepository;
+	private final ClubMemberRepository clubMemberRepository;
+	private final ClubValidator clubValidator;
 
 	@Override
 	@Transactional
 	public CreateSubmissionFormResponseDto createSubmissionForm(
-		Long clubId,
-		CreateSubmissionFormRequestDto submissionFormRequestDto
+		Long userId,
+		Long clubId, CreateSubmissionFormRequestDto submissionFormRequestDto
 	) {
 		// 1) Club 조회
 		Club findClub = clubRepository.findByIdOrElseThrow(clubId);
 
-		// 2) Club 의 JoinType 이 ProblemForm 인지 검증
+		// 2) 멤버 권한 검증
+		ClubMember findClubMember = clubMemberRepository.findByClubIdAndUserId(clubId, userId).orElseThrow(
+			() -> new BusinessException(ResponseCode.INVALID_REQUEST, "해당하는 멤버가 존재하지 않습니다.")
+		);
+		clubValidator.validateOwnerRole(findClubMember.getClubMemberRole());
+
+		// 3) Club 의 JoinType 이 SubmissionForm 인지 검증
 		if (findClub.getJoinType() != JoinType.SUBMISSION_FORM) {
-			new BusinessException(ResponseCode.INVALID_REQUEST);
+			throw new BusinessException(ResponseCode.INVALID_REQUEST, "가입방식이 `SubmissionForm`과 동일하지 않습니다.");
 		}
 
-		// 3) Form Entity 생성
+		// 4) 이미 존재하고 있다면, 생성 불가
+		if (findClub.getSubmissionForm() != null) {
+			throw new BusinessException(ResponseCode.INVALID_REQUEST, "이미 양식 존재합니다.");
+		}
+
+		// 5) Form Entity 생성
 		SubmissionForm submissionForm = new SubmissionForm(
 			submissionFormRequestDto.getSubmissionForm(),
 			findClub
 		);
 
-		// 4) 저장
+		// 6) 연관관계 주입
+		findClub.setSubmissionForm(submissionForm);
+
+		// 7) 저장
 		submissionFormRepository.save(submissionForm);
 
 		return CreateSubmissionFormResponseDto.from(submissionForm.getId());
@@ -67,17 +86,24 @@ public class SubmissionFormServiceImpl implements SubmissionFormService {
 	@Override
 	@Transactional
 	public void updateSubmissionTemplate(
+		Long userId,
 		Long clubId,
-		Long submissionId,
-		UpdateSubmissionTemplateRequestDto updateSubmissionTemplateRequestDto
+		Long submissionFormId, UpdateSubmissionTemplateRequestDto updateSubmissionTemplateRequestDto
 	) {
-		// 1) Submission Form 조회
-		SubmissionForm findSubmissionForm = submissionFormRepository.findByClubIdAndSubmissionId(clubId, submissionId)
+		// 1) 멤버 권한 검증
+		ClubMember findClubMember = clubMemberRepository.findByClubIdAndUserId(clubId, userId).orElseThrow(
+			() -> new BusinessException(ResponseCode.INVALID_REQUEST, "해당하는 멤버가 존재하지 않습니다.")
+		);
+		clubValidator.validateOwnerRole(findClubMember.getClubMemberRole());
+
+		// 2) Submission Form 조회
+		SubmissionForm findSubmissionForm = submissionFormRepository.findByClubIdAndSubmissionId(clubId,
+				submissionFormId)
 			.orElseThrow(() ->
 				new BusinessException(ResponseCode.NOT_FOUND, "Club 과 Submission Form 이 일치하는 Entity 가 존재하지 않습니다.")
 			);
 
-		// 2) Template 수정
+		// 3) Template 수정
 		if (updateSubmissionTemplateRequestDto.getSubmissionTemplate() != null) {
 			findSubmissionForm.updateSubmissionTemplate(updateSubmissionTemplateRequestDto.getSubmissionTemplate());
 		}
@@ -86,19 +112,25 @@ public class SubmissionFormServiceImpl implements SubmissionFormService {
 
 	@Override
 	@Transactional
-	public void deleteSubmissionForm(Long clubId, Long submissionId) {
-		// 1) Submission Form 조회
-		SubmissionForm findSubmissionForm = submissionFormRepository.findByClubIdAndSubmissionId(clubId, submissionId)
+	public void deleteSubmissionForm(Long userId, Long clubId, Long submissionFormId) {
+		// 1) 클럽 조회
+		Club club = clubRepository.findByIdOrElseThrow(clubId);
+
+		// 2) 멤버 권한 검증
+		ClubMember findClubMember = clubMemberRepository.findByClubIdAndUserId(clubId, userId).orElseThrow(
+			() -> new BusinessException(ResponseCode.INVALID_REQUEST, "해당하는 멤버가 존재하지 않습니다.")
+		);
+		clubValidator.validateOwnerRole(findClubMember.getClubMemberRole());
+
+		// 3) Submission Form 조회
+		SubmissionForm findSubmissionForm = submissionFormRepository.findByClubIdAndSubmissionId(clubId,
+				submissionFormId)
 			.orElseThrow(() ->
 				new BusinessException(ResponseCode.NOT_FOUND, "Club 과 Submission Form 이 일치하는 Entity 가 존재하지 않습니다.")
 			);
 
-		// 2) 삭제 - soft Delete 적용
+		// 4) 삭제
+		club.setSubmissionForm(null);
 		submissionFormRepository.delete(findSubmissionForm);
-
-		// 3) 즉시 가입 방식으로 변경
-		Club findClub = clubRepository.findByIdOrElseThrow(clubId);
-		findClub.updateJoinType(JoinType.DIRECT_JOIN);
 	}
-
 }
